fixed DLL building and Glade catalog on Windows
 * removed libtool -module flag: it should be used for dlopen libs
(gtk-vlc-player is dlopened by Glade) but had no visible impact on
building the library. also it prevented installation of the library into
the bindir!
 * explicitly specify bindir for DLL installations (it may and will be
set independently of the libdir but libtool assumes LIBDIR/../bin by
default)
 * avoid-versioning. that way the DLL name is predictable - it must be
known for the glade catalog which has been adapted to cover both the
Linux and Windows shared library names.